+++ src/main/java/com/yachaerang/yachaerangbatch/configuration/parameter/MonthlyJobParameter.java
@@ -1,4 +1,4 @@
-package com.yachaerang.yachaerangbatch.configuration.parmameter;
+package com.yachaerang.yachaerangbatch.configuration.parameter;
 
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
